requirements.txt
openai>=1.0.0
selenium>=4.0.0
beautifulsoup4>=4.9.3
requests>=2.26.0
python-dotenv>=0.19.0
langgraph>=0.0.10
pytest>=6.2.5
pytest-cov>=2.12.1
streamlit>=1.22.0
prometheus-client>=0.12.0
webdriver-manager>=3.8.0
aiohttp>=3.8.0
asyncio>=3.4.3
cachetools>=5.0.0
pydantic>=1.8.0
tenacity>=8.0.0 